requirements.txt
pronouncing
nltk
pillow
bs4